Site Manager half done

+++ WebApplication1/Models/Interventions.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace WebApplication1.Models
 {
@@ -11,8 +8,8 @@ namespace WebApplication1.Models
         public string UserId { get; set; }
         public int InterventionTypeId { get; set; }
         public int ClientId { get; set; }
-        public float InterventionCost { get; set; }
-        public float InterventionHour { get; set; }
+        public decimal InterventionCost { get; set; }
+        public decimal InterventionHour { get; set; }
         public DateTime CreateDate { get; set; }
         public int Status { get; set; }
         public int Operater { get; set; }
